requirements.txt
fastapi==0.95.1
loguru==0.7.0
pydantic==1.10.7
python-dotenv==1.0.0
torch==2.0.0
torchaudio==2.0.1
uvicorn
soundfile
numpy